criacao dos repositories

+++ files/api/src/main/java/io/github/heltonricardo/ingressoja/model/repositories/CategoriaIngressoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import io.github.heltonricardo.ingressoja.model.entities.CategoriaIngresso;
 
 public interface CategoriaIngressoRepository
-		extends PagingAndSortingRepository<CategoriaIngresso, Integer> {
+		extends PagingAndSortingRepository<CategoriaIngresso, Long> {
 }
