+++ files/api/src/main/java/io/github/heltonricardo/ingressoja/model/repositories/IngressoDisponivelRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import io.github.heltonricardo.ingressoja.model.entities.IngressoDisponivel;
 
 public interface IngressoDisponivelRepository
-		extends PagingAndSortingRepository<IngressoDisponivel, Integer> {
+		extends PagingAndSortingRepository<IngressoDisponivel, Long> {
 }
